refactor: update footer
